Organized the chaos
