+++ pages/index.tsx
@@ -2,15 +2,13 @@ import React from "react";
 
 import Link from "next/link";
 
-import Page from "../layouts/Page";
-
 export default class Index extends React.Component {
   render() {
     return (
-      <Page>
+      <div>
         <h1>Home Page!</h1>
         <Link href='/login'><a>Login</a></Link>
-      </Page>
+      </div>
     );
   }
 }
