+++ pages/login.tsx
@@ -2,15 +2,13 @@ import React from 'react';
 
 import Link from "next/link";
 
-import Page from '../layouts/Page';
-
 export default class Login extends React.Component {
     render() {
         return (
-            <Page>
+            <div>
                 <h1>Login Page!</h1>
                 <Link href='/'><a>Home</a></Link>
-            </Page>
+            </div>
         )
     }
 }
